ENH add reducer_override semantics on pickle.py
